fix (s21-libft) ft_strlcpy | ft_strcat

+++ school21/42cursus/libft/ft_strlen.c
@@ -1,12 +1,12 @@
+#include "libft.h"
 
-int ft_strlen(const char *s)
+size_t ft_strlen(const char *s)
 {
-    int len;
+    size_t len;
 
     len = 0;
     while (s[len] != '\0') {
         len += 1;
     }
-
     return len;
 }
